policy_service/requirements.txt
Django==4.2
djangorestframework==3.15
djangorestframework-simplejwt==5.2.2
kafka-python==2.1.0
python-dotenv==1.0.0
django-filter==24.3